pyproject.toml
[tool.poetry]
name = "inconfluential"
version = "0.1.0"
description = ""
authors = ["dinjou <spencerklay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.3"
setuptools = "^75.4.0"
python-dotenv = "^1.0.1"
atlassian-python-api = "^3.41.16"
markdownify = "^0.13.1"
tqdm = "^4.67.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
